func: port to riscv refactor: counter funcs and makefile  - move ISA dependent counter func definition to counter.h  - move impls to counter_$isa.c  - Usage: make ARCH=<alpha,x86,riscv>
