+++ clock.h
@@ -1,12 +1,5 @@
 /* Routines for using cycle counter */
 
-/* Start the counter */
-void start_counter();
-
-/* Get # cycles since counter started */
-double get_counter();
-
-
 /* Measure overhead for counter */
 double ovhd();
