+++ counter.h
@@ -1,0 +1,8 @@
+/* API for counters*/
+
+/* Start the counter */
+void start_counter();
+
+/* Get # cycles since counter started */
+double get_counter();
+
