+++ counter_x86.c
@@ -1,0 +1,45 @@
+/* Use X86 cycle timer to compute cycles */
+#if defined(__i386__) || defined(__x86_64__)
+
+#include <stdio.h>
+#include "counter.h"
+
+/* Keep track of most recent reading of cycle counter */
+static unsigned cyc_hi = 0;
+static unsigned cyc_lo = 0;
+
+void access_counter(unsigned *hi, unsigned *lo)
+{
+  /* Get cycle counter */
+  asm("rdtsc; movl %%edx,%0; movl %%eax,%1"
+      : "=r" (*hi), "=r" (*lo)
+      : /* No input */
+      : "%edx", "%eax");
+}
+
+void start_counter()
+{
+  access_counter(&cyc_hi, &cyc_lo);
+}
+
+double get_counter()
+{
+  unsigned ncyc_hi, ncyc_lo;
+  unsigned hi, lo, borrow;
+  double result;
+  /* Get cycle counter */
+  access_counter(&ncyc_hi, &ncyc_lo);
+  /* Do double precision subtraction */
+  lo = ncyc_lo - cyc_lo;
+  borrow = lo > ncyc_lo;
+  hi = ncyc_hi - cyc_hi - borrow;
+  result = (double) hi * (1 << 30) * 4 + lo;
+  if (result < 0) {
+    fprintf(stderr, "Error: Cycle counter returning negative value: %.0f\n", result);
+  }
+  return result;
+}
+
+#else
+#error "Macro `__i386__` or `__x86_64__` wasn't defined!"
+#endif /* __i386__ */
